모니터 기능 분리
