+++ cogs/cog_general/__init__.py
@@ -1,10 +1,8 @@
-import os
 import platform
 from datetime import datetime
 from time import time
 
 import discord
-from discord import app_commands
 from discord.ext import commands
 from discord.ext.commands import Context
 
@@ -125,117 +123,6 @@ class General(commands.Cog, name="general"):
             await context.send("개인 메시지로 초대 링크를 보냈어요! 📩", ephemeral=True)
         except discord.Forbidden:
             await context.send(embed=embed, ephemeral=True)
-
-    @commands.hybrid_command(
-        name="restrict",
-        description="봇의 소스 코드를 보여줍니다.",
-    )
-    @app_commands.choices(
-        onoff=[
-            app_commands.Choice(name="on", value="on"),
-            app_commands.Choice(name="off", value="off"),
-        ]
-    )
-    async def restrict(self, context: Context, onoff: str) -> None:
-        if onoff == "on":
-            self._restrict = True
-            await context.send(
-                f"{os.getenv('PJY_NAME')} 몰컴 감시기 가동", ephemeral=True
-            )
-        elif onoff == "off":
-            self._restrict = False
-            await context.send(
-                f"{os.getenv('PJY_NAME')} 몰컴 감시기 중지", ephemeral=True
-            )
-        else:
-            await context.send("잘못된 입력입니다.")
-
-    @commands.Cog.listener()
-    async def on_presence_update(self, before: discord.Member, after: discord.Member):
-        guild = before.guild
-        first_channel = guild.text_channels[0]
-        test_channel = guild.text_channels[-1]
-
-        if guild.id == int(os.getenv("PJY_GUILD_ID")):
-            self.logger.info(f"user: {after.display_name}")
-            self.logger.info(f"{before.activity} -> {after.activity}")
-            if after.id == int(os.getenv("PJY_ID")):
-                self.logger.info("타켓 확인")
-                if (
-                    before.desktop_status != after.desktop_status
-                    and after.desktop_status == discord.Status.online
-                ):
-                    self.logger.info(f"{os.getenv('PJY_NAME')} 컴퓨터 온라인 검거")
-                    if self._restrict:
-                        await test_channel.send(
-                            f"{os.getenv('PJY_NAME')} 컴퓨터 온라인 검거.", silent=True
-                        )
-
-                before_activity = (
-                    "" if before.activity is None else (before.activity.name or "")
-                )
-                after_activity = (
-                    "" if after.activity is None else (after.activity.name or "")
-                )
-                if before_activity != after_activity:
-                    if "VALORANT" in after_activity:
-                        if after.voice is None:
-                            self.logger.info(f"{os.getenv('PJY_NAME')} 몰로란트 시작")
-                            if self._restrict:
-                                self._start_time_val = datetime.now()
-                                await first_channel.send(
-                                    "WARNING! WARNING! WARNING!\n"
-                                    f"{after.mention}몰로란트 검거\n"
-                                    "WARNING! WARNING! WARNING!"
-                                )
-                    elif "League of Legends" in after_activity:
-                        if after.voice is None:
-                            self.logger.info(f"{os.getenv('PJY_NAME')} 몰롤 시작")
-                            if self._restrict:
-                                self._start_time_lol = datetime.now()
-                                await first_channel.send(
-                                    "WARNING! WARNING! WARNING!\n"
-                                    f"{after.mention}몰롤 검거\n"
-                                    "WARNING! WARNING! WARNING!"
-                                )
-                    elif "VALORANT" in before_activity and after_activity == "":
-                        if after.voice is None:
-                            self.logger.info(f"{os.getenv('PJY_NAME')} 게임 종료")
-                            if self._restrict:
-                                _time = datetime.now() - self._start_time_val
-                                seconds = _time.seconds
-                                minute = seconds // 60
-                                time_format = "몰컴 시간: "
-                                if minute:
-                                    time_format += f"{minute}분 "
-                                time_format += f"{seconds % 60}초"
-
-                                await first_channel.send(
-                                    "WARNING! WARNING! WARNING!\n"
-                                    f"{after.mention}몰로란트 종료\n"
-                                    f"{time_format}\n"
-                                    "WARNING! WARNING! WARNING!"
-                                )
-                    elif (
-                        "League of Legends" in before_activity and after_activity == ""
-                    ):
-                        if after.voice is None:
-                            self.logger.info(f"{os.getenv('PJY_NAME')} 게임 종료")
-                            if self._restrict:
-                                _time = datetime.now() - self._start_time_lol
-                                seconds = _time.seconds
-                                minute = seconds // 60
-                                time_format = "몰컴 시간: "
-                                if minute:
-                                    time_format += f"{minute}분 "
-                                time_format += f"{seconds % 60}초"
-
-                                await first_channel.send(
-                                    "WARNING! WARNING! WARNING!\n"
-                                    f"{after.mention}몰롤 종료\n"
-                                    f"{time_format}\n"
-                                    "WARNING! WARNING! WARNING!"
-                                )
 
 
 async def setup(bot: ServantBot) -> None:
